- Feature: Copy configuration values from existing IXMLOptions
- Feature: Let the caller of the Serializer decide whether to use Attributes or Text-Only Children
- Bugfix: Enforce Serialization order for children and attributes
- Minor fixes and code cleanup
- Updated tests

+++ MapXML/Behaviors/XMLMemberBehavior.cs
@@ -13,7 +13,7 @@ namespace MapXML.Behaviors
         //-----------------------------------------------------------------//
         public string NodeName => _attribute?.NodeName ?? this.Member.Name;
         public abstract Type TypeToCreate { get; }
-        public int SerializationOrder { get; set; } = int.MaxValue;
+        public int SerializationOrder { get; set; } = byte.MaxValue;
         public XMLSourceType SourceType { get; }
         internal DeserializationPolicy Policy { get; }
         private ConvertFromString? _conversionFromString_nullable;
